refactor: Frame class moved to io.py
tests: io tests passing
