+++ src/cala/streaming/nodes/iter/detect.py
@@ -12,7 +12,7 @@ from skimage.restoration import estimate_sigma
 from sklearn.decomposition import NMF
 from sklearn.feature_extraction.image import PatchExtractor
 
-from cala.config.pipe import Frame
+from cala.io import Frame
 from cala.streaming.core import Axis, Component, Parameters
 from cala.streaming.stores.common import Footprints, Traces
 from cala.streaming.stores.odl import ComponentStats, Overlaps, PixelStats, Residuals
