+++ src/cala/streaming/nodes/iter/footprints.py
@@ -7,7 +7,7 @@ import xarray as xr
 from river.base import SupervisedTransformer
 from sklearn.exceptions import NotFittedError
 
-from cala.config.pipe import Frame
+from cala.io import Frame
 from cala.streaming.core import Axis, Parameters
 from cala.streaming.stores.common import Footprints
 from cala.streaming.stores.odl import ComponentStats, PixelStats
